Add files via upload
Sorry, it has to do with the dependencies I accidentally forgot to acknowledge in the code. I used the p5.vscode extension. I am not sure to what extent we are allowed to modify the json file, but running the command npm install p5 --save will automatically add the p5 dependency inside the json file. 

Otherwise, I have modified my submission to include the corresponding missing files, which are two libraries and the corresponding link between the html file and my submission. Hopefully, I can still participate in the contest.